tests for stats extraction and total elapsed time; updates #78'
