+++ source/consumers/add_ons/filter_block.cpp
