+++ tests/consumers/add_ons/periodic_trgger.cpp
@@ -1,0 +1,8 @@
+#include <gtest/gtest.h>
+#include "periodic_trigger.h"
+
+using namespace DAQuiri;
+
+TEST(PeriodicFilter, Init)
+{
+}
